Changes the structure of --install by using symbolic links instead of file copying. This makes file maintance easier by altering them directly, instead of changing on the repository to then, install the new changes.
With that, the old update function has been removed.

This history might look ugly, but this is due to the changes being made a few weeks ago in my dotfile repo.
Then, I just copied a new dotmanager.pl file instead of applying changes to the current version. As this
is a personal repo, it does not matter much.